# bindings/src/types/tari-indexer-client/TransactionEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PrunedTransaction } from "../PrunedTransaction";
import type { TransactionId } from "../TransactionId";
import type { TransactionResultSummary } from "./TransactionResultSummary";
import type { TransactionSource } from "./TransactionSource";

export type TransactionEntry = {
  transaction_id: TransactionId;
  /**
   * Pruned transaction — blob commitments retained, raw blob bytes omitted to keep the
   * response size bounded. The transaction id and signatures remain verifiable.
   */
  transaction: PrunedTransaction;
  created_at: string;
  /**
   * Result summary from the locally indexed transaction receipt. None when no receipt has been
   * indexed yet (pending or aborted).
   */
  summary: TransactionResultSummary | null;
  /**
   * Reason the transaction was rejected by mempool validation when it was submitted through
   * this indexer. None if the transaction was not rejected at submission.
   */
  rejected_reason: string | null;
  /**
   * Where this indexer learned of the transaction.
   */
  source: TransactionSource;
};