# bindings/src/types/tari-indexer-client/TransactionSource.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Where an indexer learned of a transaction it has stored.
 */
export type TransactionSource = "local" | "gossip";