# docs/developer-docs/public/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tari_indexer",
    "description": "The Tari indexer implementation",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.39.2"
  },
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": [
          "handlers::transactions"
        ],
        "description": "Submit a transaction to validators responsible for the involved shards",
        "operationId": "submit_transaction",
        "responses": {
          "200": {
            "description": "Transaction submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction or request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to submit transaction due to an internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "All validators failed to process the transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/dry-run": {
      "post": {
        "tags": [
          "handlers::transactions"
        ],
        "description": "Submit a transaction as a dry-run",
        "operationId": "submit_transaction_dry_run",
        "responses": {
          "200": {
            "description": "Dry-run transaction processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitTransactionDryRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction or request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to process dry-run transaction due to an internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/transactions/recent": {
      "get": {
        "tags": [
          "handlers::transactions"
        ],
        "description": "List recent transactions",
        "operationId": "list_recent_transactions",
        "responses": {
          "200": {
            "description": "List of recent transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRecentTransactionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list recent transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "handlers::transactions"
        ],
        "description": "Get a transaction (including its instructions, fee instructions and signatures) by transaction ID.\n\nNOTE: This reads the indexer's local transaction store, which holds transactions submitted through this indexer and — when gossip indexing is enabled, see `index_gossiped_transactions` on /info — those observed on the network transaction topic. A transaction the indexer never saw, or one that has aged past its retention window, returns 404 here even after it has been finalized. This differs from transaction receipts (see the /transaction-receipts endpoints), which are synced from the network. For the execution result of any transaction regardless of where it was submitted, use /transactions/{transaction_id}/result.",
        "operationId": "get_transaction",
        "responses": {
          "200": {
            "description": "Transaction found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not known to this indexer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}/result": {
      "get": {
        "tags": [
          "handlers::transactions"
        ],
        "description": "Get the result of a submitted transaction (by transaction ID)",
        "operationId": "get_transaction_result",
        "responses": {
          "200": {
            "description": "Transaction result found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionResultResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaction result not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch transaction result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetTransactionResponse": {
        "type": "object",
        "required": [
          "transaction"
        ],
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TransactionEntry",
            "description": "The stored transaction, including its instructions, fee instructions and signatures."
          }
        }
      },
      "GetTransactionResultResponse": {
        "type": "object",
        "required": [
          "result"
        ],
        "properties": {
          "result": {
            "type": "object",
            "description": "The result of the transaction, which may be pending (not yet finalized) or finalized with details such as the\nfinal decision, execution result, and timestamps"
          }
        }
      },
...
      "ListRecentTransactionsResponse": {
        "type": "object",
        "required": [
          "transactions"
        ],
        "properties": {
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionEntry"
            }
          }
        }
      },
...
      "SubmitTransactionDryRunResponse": {
        "type": "object",
        "required": [
          "transaction_id",
          "result"
        ],
        "properties": {
          "result": {
            "type": "object",
            "description": "The result of the dry-run execution, including any emitted events and state changes, but without a final\ndecision or commitment to the ledger"
          },
          "transaction_id": {
            "type": "string",
            "description": "The ID of the transaction that was dry-run"
          }
        }
      },
      "SubmitTransactionResponse": {
        "type": "object",
        "required": [
          "transaction_id"
        ],
        "properties": {
          "transaction_id": {
            "type": "string",
            "description": "The ID of the submitted transaction"
          }
        }
      },
...
      "TransactionEntry": {
        "type": "object",
        "required": [
          "transaction_id",
          "transaction",
          "created_at",
          "source"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "rejected_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Reason the transaction was rejected by mempool validation when it was submitted through\nthis indexer. None if the transaction was not rejected at submission."
          },
          "source": {
            "$ref": "#/components/schemas/TransactionSource",
            "description": "Where this indexer learned of the transaction."
          },
          "summary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransactionResultSummary",
                "description": "Result summary from the locally indexed transaction receipt. None when no receipt has been\nindexed yet (pending or aborted)."
              }
            ]
          },
          "transaction": {
            "type": "object",
            "description": "Pruned transaction — blob commitments retained, raw blob bytes omitted to keep the\nresponse size bounded. The transaction id and signatures remain verifiable."
          },
          "transaction_id": {
            "type": "string"
          }
        }
      },
      "TransactionResultSummary": {
        "type": "object",
        "required": [
          "outcome",
          "total_fees_paid",
          "finalized_at"
        ],
        "properties": {
          "finalized_at": {
            "type": "string",
            "format": "date-time"
          },
          "outcome": {
            "type": "string"
          },
          "total_fees_paid": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "TransactionSource": {
        "type": "string",
        "description": "Where an indexer learned of a transaction it has stored.",
        "enum": [
          "local",
          "gossip"
        ]
      },
...
    }
  }
}